rest-api/src/main/java/com/example/restapi/api/v1/mapper/CustomerMapper.java
package com.example.restapi.api.v1.mapper;

import org.mapstruct.Mapper;
import org.mapstruct.factory.Mappers;

import com.example.restapi.api.v1.model.CustomerDTO;
import com.example.restapi.domain.Customer;

/**
 * Data transfer object mapper using MapStruct
 */
@Mapper
public interface CustomerMapper {

    CustomerMapper INSTANCE = Mappers.getMapper(CustomerMapper.class);

    CustomerDTO customerToCustomerDTO(Customer customer);
    Customer customerDTOToCustomer(CustomerDTO customerDTO);
}
